+++ packages/fury-adapter-oas3-parser/test/integration/components-test.js
@@ -19,6 +19,11 @@ describe('components', () => {
     return testParseFixture(file);
   });
 
+  it("'Media Type Object' examples references", () => {
+    const file = path.join(fixtures, 'media-type-object-examples');
+    return testParseFixture(file);
+  });
+
   it("'Responses Object' response references", () => {
     const file = path.join(fixtures, 'responses-object-response');
     return testParseFixture(file);